# docs/integrations/ownlevel-chatgpt-action.openapi.yaml
openapi: 3.1.0
info:
  title: OWNLEVEL Nutrition
  version: 1.2.0
  description: >-
    Comprueba una clave privada y registra comidas estructuradas en la cuenta
    OWNLEVEL vinculada. No ofrece operaciones de lectura nutricional.
servers:
  - url: https://www.ownlevel.fit
paths:
  /api/integrations/chatgpt/status:
    get:
      operationId: checkConnection
      summary: Comprobar la conexión con OWNLEVEL
      description: >-
        Valida la clave y su permiso meals:write sin leer ni devolver datos
        personales o nutricionales.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conexión autenticada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionSuccess'
        '401':
          description: Clave ausente, inválida o revocada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error interno sin detalles sensibles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/integrations/chatgpt/meals:
    post:
      operationId: logMeal
      summary: Registrar una comida en OWNLEVEL
      description: >-
        Registra una comida en la fecha indicada. Omití un macro cuando no pueda
        estimarse: null significa desconocido y 0 significa conocido y realmente cero.
        No inventes valores. Conservá la misma idempotency_key al reintentar una acción.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MealInput'
      responses:
        '200':
          description: Comida creada o retry idempotente resuelto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealSuccess'
        '400':
          description: Payload inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Clave ausente, inválida o revocada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >-
            Posible duplicado humano. Confirmá con el usuario y repetí con una nueva
            idempotency_key y force_duplicate=true sólo si realmente fue otra comida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Body mayor al límite permitido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error interno sin detalles sensibles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OWNLEVEL token
  schemas:
    ConnectionSuccess:
      type: object
      additionalProperties: false
      required: [ok, connected]
      properties:
        ok: { type: boolean, const: true }
        connected: { type: boolean, const: true }
    MealInput:
      type: object
      additionalProperties: false
      required: [title, calories, idempotency_key]
      properties:
        date:
          type: string
          format: date
          description: Fecha lógica en Córdoba. Si se omite, OWNLEVEL usa hoy en America/Argentina/Cordoba.
        title:
          type: string
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 2000
        calories:
          type: integer
          minimum: 1
        protein_g:
          type: [number, 'null']
          minimum: 0
        carbs_g:
          type: [number, 'null']
          minimum: 0
        fat_g:
          type: [number, 'null']
          minimum: 0
        idempotency_key:
          type: string
          maxLength: 200
          description: Identificador único y estable para este intento lógico.
        force_duplicate:
          type: boolean
          default: false
    MealSuccess:
      type: object
      additionalProperties: false
      required: [ok, created, idempotent_replay, meal]
      properties:
        ok: { type: boolean, const: true }
        created: { type: boolean }
        idempotent_replay: { type: boolean }
        meal:
          type: object
          additionalProperties: false
          required: [id, date, title, calories, protein_g, carbs_g, fat_g]
          properties:
            id: { type: string, format: uuid }
            date: { type: string, format: date }
            title: { type: [string, 'null'] }
            calories: { type: integer }
            protein_g: { type: [number, 'null'] }
            carbs_g: { type: [number, 'null'] }
            fat_g: { type: [number, 'null'] }
    Error:
      type: object
      additionalProperties: false
      required: [ok, error, message]
      properties:
        ok: { type: boolean, const: false }
        error:
          type: string
          enum: [invalid_request, invalid_token, possible_duplicate, internal_error]
        message: { type: string }